RPA: Move rpa_manager from electroncash to electroncash.rpa
